Refs #749: Updating xforms-server library.

+++ motech-mobile-webapp/src/main/java/org/motechproject/mobile/web/OXDFormUploadServlet.java
@@ -202,7 +202,7 @@ public class OXDFormUploadServlet implements ApplicationContextAware {
 					if (!impService.getFormProcessSuccess().equalsIgnoreCase(
 							studyForms[s][f])) {
 						dataOutput.writeByte((byte) s);
-						dataOutput.writeByte((byte) f);
+						dataOutput.writeShort((short) f);
 						dataOutput.writeUTF(studyForms[s][f]);
 					}
 				}
